Avoid stack overflow when disposing batched sink
