+++ test/Serilog.Sinks.AwsCloudWatch.Tests/CloudWatchLogsSinkTests.cs
@@ -36,7 +36,7 @@ namespace Serilog.Sinks.AwsCloudWatch.Tests
             var textFormatterMock = new Mock<ITextFormatter>(MockBehavior.Strict);
             textFormatterMock.Setup(s => s.Format(It.IsAny<LogEvent>(), It.IsAny<TextWriter>())).Callback((LogEvent l, TextWriter t) => l.RenderMessage(t));
             var options = new CloudWatchSinkOptions { TextFormatter = textFormatterMock.Object, LogGroupName = "Test-Log-Group-Name" };
-            var sink = new CloudWatchLogSink(client.Object, options);
+            var sink = new CloudWatchLogsBatchedSink(client.Object, options);
             var events = Enumerable.Range(0, 10)
                 .Select(_ => // create 10 events with message length of 12
                     new LogEvent(
@@ -93,7 +93,7 @@ namespace Serilog.Sinks.AwsCloudWatch.Tests
             var textFormatterMock = new Mock<ITextFormatter>(MockBehavior.Strict);
             textFormatterMock.Setup(s => s.Format(It.IsAny<LogEvent>(), It.IsAny<TextWriter>())).Callback((LogEvent l, TextWriter t) => l.RenderMessage(t));
             var options = new CloudWatchSinkOptions { LogGroupName = Guid.NewGuid().ToString(), TextFormatter = textFormatterMock.Object };
-            var sink = new CloudWatchLogSink(client.Object, options);
+            var sink = new CloudWatchLogsBatchedSink(client.Object, options);
             var events = Enumerable.Range(0, 10)
                 .Select(_ => // create 10 events with message length of 12
                     new LogEvent(
@@ -157,7 +157,7 @@ namespace Serilog.Sinks.AwsCloudWatch.Tests
             var textFormatterMock = new Mock<ITextFormatter>(MockBehavior.Strict);
             textFormatterMock.Setup(s => s.Format(It.IsAny<LogEvent>(), It.IsAny<TextWriter>())).Callback((LogEvent l, TextWriter t) => l.RenderMessage(t));
             var options = new CloudWatchSinkOptions { CreateLogGroup = false, TextFormatter = textFormatterMock.Object, LogGroupName = "Test-Log-Group-Name" };
-            var sink = new CloudWatchLogSink(client.Object, options);
+            var sink = new CloudWatchLogsBatchedSink(client.Object, options);
             var events = Enumerable.Range(0, 10)
                 .Select(_ => // create 10 events with message length of 12
                     new LogEvent(
@@ -208,7 +208,7 @@ namespace Serilog.Sinks.AwsCloudWatch.Tests
             var textFormatterMock = new Mock<ITextFormatter>(MockBehavior.Strict);
             textFormatterMock.Setup(s => s.Format(It.IsAny<LogEvent>(), It.IsAny<TextWriter>())).Callback((LogEvent l, TextWriter t) => l.RenderMessage(t));
             var options = new CloudWatchSinkOptions { LogGroupName = Guid.NewGuid().ToString(), LogStreamNameProvider = new NonUniqueLogStreamNameProvider(), TextFormatter = textFormatterMock.Object };
-            var sink = new CloudWatchLogSink(client.Object, options);
+            var sink = new CloudWatchLogsBatchedSink(client.Object, options);
             var events = Enumerable.Range(0, 10)
                 .Select(_ => // create 10 events with message length of 12
                     new LogEvent(
@@ -280,7 +280,7 @@ namespace Serilog.Sinks.AwsCloudWatch.Tests
             var textFormatterMock = new Mock<ITextFormatter>(MockBehavior.Strict);
             textFormatterMock.Setup(s => s.Format(It.IsAny<LogEvent>(), It.IsAny<TextWriter>())).Callback((LogEvent l, TextWriter t) => l.RenderMessage(t));
             var options = new CloudWatchSinkOptions { TextFormatter = textFormatterMock.Object, LogGroupName = "Test-Log-Group-Name" };
-            var sink = new CloudWatchLogSink(client.Object, options);
+            var sink = new CloudWatchLogsBatchedSink(client.Object, options);
             var largeEventMessage = CreateMessage(CloudWatchLogSink.MaxLogEventSize + 1);
             var events = new LogEvent[]
             {
@@ -335,7 +335,7 @@ namespace Serilog.Sinks.AwsCloudWatch.Tests
             var textFormatterMock = new Mock<ITextFormatter>(MockBehavior.Strict);
             textFormatterMock.Setup(s => s.Format(It.IsAny<LogEvent>(), It.IsAny<TextWriter>())).Callback((LogEvent l, TextWriter t) => l.RenderMessage(t));
             var options = new CloudWatchSinkOptions { TextFormatter = textFormatterMock.Object, LogGroupName = "Test-Log-Group-Name" };
-            var sink = new CloudWatchLogSink(client.Object, options);
+            var sink = new CloudWatchLogsBatchedSink(client.Object, options);
             var events = Enumerable.Range(0, 20)
                 .Select(i => // create multipe events with message length of 12
                     new LogEvent(
@@ -407,7 +407,7 @@ namespace Serilog.Sinks.AwsCloudWatch.Tests
             var textFormatterMock = new Mock<ITextFormatter>(MockBehavior.Strict);
             textFormatterMock.Setup(s => s.Format(It.IsAny<LogEvent>(), It.IsAny<TextWriter>())).Callback((LogEvent l, TextWriter t) => l.RenderMessage(t));
             var options = new CloudWatchSinkOptions { TextFormatter = textFormatterMock.Object, LogGroupName = "Test-Log-Group-Name" };
-            var sink = new CloudWatchLogSink(client.Object, options);
+            var sink = new CloudWatchLogsBatchedSink(client.Object, options);
             var events = Enumerable.Range(0, CloudWatchLogSink.MaxLogEventBatchCount + 1)
                 .Select(i =>
                     new LogEvent(
@@ -474,7 +474,7 @@ namespace Serilog.Sinks.AwsCloudWatch.Tests
             var textFormatterMock = new Mock<ITextFormatter>(MockBehavior.Strict);
             textFormatterMock.Setup(s => s.Format(It.IsAny<LogEvent>(), It.IsAny<TextWriter>())).Callback((LogEvent l, TextWriter t) => l.RenderMessage(t));
             var options = new CloudWatchSinkOptions { TextFormatter = textFormatterMock.Object, LogGroupName = "Test-Log-Group-Name" };
-            var sink = new CloudWatchLogSink(client.Object, options);
+            var sink = new CloudWatchLogsBatchedSink(client.Object, options);
             var events = Enumerable.Range(0, 256) // 256 4 KB messages matches our max batch size, but we want to test a "less nice" scenario, so we'll create 256 5 KB messages
                 .Select(i =>
                     new LogEvent(
@@ -541,7 +541,7 @@ namespace Serilog.Sinks.AwsCloudWatch.Tests
             var textFormatterMock = new Mock<ITextFormatter>(MockBehavior.Strict);
             textFormatterMock.Setup(s => s.Format(It.IsAny<LogEvent>(), It.IsAny<TextWriter>())).Callback((LogEvent l, TextWriter t) => l.RenderMessage(t));
             var options = new CloudWatchSinkOptions { TextFormatter = textFormatterMock.Object, LogGroupName = "Test-Log-Group-Name" };
-            var sink = new CloudWatchLogSink(client.Object, options);
+            var sink = new CloudWatchLogsBatchedSink(client.Object, options);
             var events = Enumerable.Range(0, 10)
                 .Select(_ => // create 10 events with message length of 12
                     new LogEvent(
@@ -594,7 +594,7 @@ namespace Serilog.Sinks.AwsCloudWatch.Tests
             var textFormatterMock = new Mock<ITextFormatter>(MockBehavior.Strict);
             textFormatterMock.Setup(s => s.Format(It.IsAny<LogEvent>(), It.IsAny<TextWriter>())).Callback((LogEvent l, TextWriter t) => l.RenderMessage(t));
             var options = new CloudWatchSinkOptions { TextFormatter = textFormatterMock.Object, LogGroupName = "Test-Log-Group-Name" };
-            var sink = new CloudWatchLogSink(client.Object, options);
+            var sink = new CloudWatchLogsBatchedSink(client.Object, options);
             var events = Enumerable.Range(0, 10)
                 .Select(_ => // create 10 events with message length of 12
                     new LogEvent(
@@ -636,7 +636,7 @@ namespace Serilog.Sinks.AwsCloudWatch.Tests
             var textFormatterMock = new Mock<ITextFormatter>(MockBehavior.Strict);
             textFormatterMock.Setup(s => s.Format(It.IsAny<LogEvent>(), It.IsAny<TextWriter>())).Callback((LogEvent l, TextWriter t) => l.RenderMessage(t));
             var options = new CloudWatchSinkOptions { TextFormatter = textFormatterMock.Object, LogGroupName = "Test-Log-Group-Name" };
-            var sink = new CloudWatchLogSink(client.Object, options);
+            var sink = new CloudWatchLogsBatchedSink(client.Object, options);
             var events = Enumerable.Range(0, 10)
                 .Select(_ => // create 10 events with message length of 12
                     new LogEvent(
@@ -686,7 +686,7 @@ namespace Serilog.Sinks.AwsCloudWatch.Tests
             var textFormatterMock = new Mock<ITextFormatter>(MockBehavior.Strict);
             textFormatterMock.Setup(s => s.Format(It.IsAny<LogEvent>(), It.IsAny<TextWriter>())).Callback((LogEvent l, TextWriter t) => l.RenderMessage(t));
             var options = new CloudWatchSinkOptions { TextFormatter = textFormatterMock.Object, LogGroupName = "Test-Log-Group-Name" };
-            var sink = new CloudWatchLogSink(client.Object, options);
+            var sink = new CloudWatchLogsBatchedSink(client.Object, options);
             var events = Enumerable.Range(0, 10)
                 .Select(_ => // create 10 events with message length of 12
                     new LogEvent(
@@ -732,7 +732,7 @@ namespace Serilog.Sinks.AwsCloudWatch.Tests
             var textFormatterMock = new Mock<ITextFormatter>(MockBehavior.Strict);
             textFormatterMock.Setup(s => s.Format(It.IsAny<LogEvent>(), It.IsAny<TextWriter>())).Callback((LogEvent l, TextWriter t) => l.RenderMessage(t));
             var options = new CloudWatchSinkOptions { TextFormatter = textFormatterMock.Object, LogGroupName = "Test-Log-Group-Name" };
-            var sink = new CloudWatchLogSink(client.Object, options);
+            var sink = new CloudWatchLogsBatchedSink(client.Object, options);
             var events = Enumerable.Range(0, 10)
                 .Select(_ => // create 10 events with message length of 12
                     new LogEvent(
@@ -776,7 +776,7 @@ namespace Serilog.Sinks.AwsCloudWatch.Tests
             var textFormatterMock = new Mock<ITextFormatter>(MockBehavior.Strict);
             textFormatterMock.Setup(s => s.Format(It.IsAny<LogEvent>(), It.IsAny<TextWriter>())).Callback((LogEvent l, TextWriter t) => l.RenderMessage(t));
             var options = new CloudWatchSinkOptions { TextFormatter = textFormatterMock.Object, LogGroupName = "Test-Log-Group-Name" };
-            var sink = new CloudWatchLogSink(client.Object, options);
+            var sink = new CloudWatchLogsBatchedSink(client.Object, options);
             var events = Enumerable.Range(0, 10)
                 .Select(_ => // create 10 events with message length of 12
                     new LogEvent(
@@ -828,7 +828,7 @@ namespace Serilog.Sinks.AwsCloudWatch.Tests
             var textFormatterMock = new Mock<ITextFormatter>(MockBehavior.Strict);
             textFormatterMock.Setup(s => s.Format(It.IsAny<LogEvent>(), It.IsAny<TextWriter>())).Callback((LogEvent l, TextWriter t) => l.RenderMessage(t));
             var options = new CloudWatchSinkOptions { LogStreamNameProvider = logStreamNameProvider.Object, TextFormatter = textFormatterMock.Object, LogGroupName = "Test-Log-Group-Name" };
-            var sink = new CloudWatchLogSink(client.Object, options);
+            var sink = new CloudWatchLogsBatchedSink(client.Object, options);
             var events = Enumerable.Range(0, 10)
                 .Select(_ => // create 10 events with message length of 12
                     new LogEvent(
@@ -882,7 +882,7 @@ namespace Serilog.Sinks.AwsCloudWatch.Tests
             var textFormatterMock = new Mock<ITextFormatter>(MockBehavior.Strict);
             textFormatterMock.Setup(s => s.Format(It.IsAny<LogEvent>(), It.IsAny<TextWriter>())).Callback((LogEvent l, TextWriter t) => l.RenderMessage(t));
             var options = new CloudWatchSinkOptions { TextFormatter = textFormatterMock.Object, LogGroupName = "Test-Log-Group-Name" };
-            var sink = new CloudWatchLogSink(client.Object, options);
+            var sink = new CloudWatchLogsBatchedSink(client.Object, options);
             var events = Enumerable.Range(0, 10)
                 .Select(_ => // create 10 events with message length of 12
                     new LogEvent(
